# frontend/src/api/generated/current-openapi.yaml
openapi: 3.1.0
info:
  title: ShuttleCube Internal Management API
  version: 0.1.0
  description: |
    羽毛球培训与场地经营管理系统的同源内部接口契约。
    该接口仅供已登录管理界面使用，不是外部公共 API。
    `/agents`、`/workflows`、`/agent-runs` 和 `/approval-requests` 为未来 Agent 契约草案，
    当前版本不得实现、生成客户端调用或作为验收范围。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /courts:
    get:
      tags:
        - Directory
      summary: 查询场地
      operationId: listCourts
      responses:
        "200":
          description: 场地列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Court"
    post:
      tags:
        - Directory
      summary: 新增场地
      operationId: createCourt
      requestBody:
        $ref: "#/components/requestBodies/CourtWrite"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Court"
        "422":
          $ref: "#/components/responses/ValidationError"
  /coaches:
    get:
      tags:
        - Directory
      summary: 查询教练
      operationId: listCoaches
      responses:
        "200":
          description: 教练列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coach"
    post:
      tags:
        - Directory
      summary: 新增教练
      operationId: createCoach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachWrite"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coach"
        "422":
          $ref: "#/components/responses/ValidationError"
  /coaches/{coach_id}:
    put:
      tags:
        - Directory
      summary: 编辑教练资料
      operationId: updateCoach
      parameters:
        - name: coach_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachUpdate"
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coach"
        "409":
          $ref: "#/components/responses/ConcurrentChange"
  /coaches/{coach_id}/rates:
    get:
      tags:
        - Directory
      summary: 查询教练固定班和私教费用标准历史
      operationId: listCoachRates
      parameters:
        - name: coach_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 按业务类型和生效日期倒序返回费用标准
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachRate"
        "404":
          description: 教练不存在
  /venue/settings:
    get:
      tags:
        - Directory
      summary: 获取机构营业设置
      operationId: getVenueSettings
      responses:
        "200":
          description: 当前设置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueSettings"
    put:
      tags:
        - Directory
      summary: 更新机构营业设置
      operationId: updateVenueSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueSettingsWrite"
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueSettings"
        "409":
          $ref: "#/components/responses/ConcurrentChange"
  /students:
    get:
      tags:
        - Customers
      summary: 查询学员
      operationId: listStudents
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: 分页学员列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentPage"
    post:
      tags:
        - Customers
      summary: 新增学员及家长关联
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentWrite"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shuttlecube_session
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Search:
      name: q
      in: query
      schema:
        type: string
        maxLength: 100
...
  requestBodies:
    CourtWrite:
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - code
              - name
            properties:
              code:
                type: string
                minLength: 1
                maxLength: 20
              name:
                type: string
                minLength: 1
                maxLength: 100
              notes:
                type:
                  - string
                  - "null"
                maxLength: 1000
  responses:
...
    ValidationError:
      description: 输入或业务规则校验失败
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblem"
    ConcurrentChange:
      description: 记录版本已变化或目标事实已被其他操作处理
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    IdentifierVersion:
      type: object
      required:
        - id
        - version
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 1
    Money:
      type: string
      pattern: ^\d+(\.\d{1,2})?$
      examples:
        - "120.00"
    Problem:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                required:
                  - field
                  - message
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    Court:
      allOf:
        - $ref: "#/components/schemas/IdentifierVersion"
        - type: object
          required:
            - code
            - name
            - status
          properties:
            code:
              type: string
            name:
              type: string
            status:
              type: string
              enum:
                - active
                - inactive
            notes:
              type:
                - string
                - "null"
    CoachWrite:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type:
            - string
            - "null"
        teaching_tags:
          type: array
          items:
            type: string
        notes:
          type:
            - string
            - "null"
        fixed_class_fee:
          anyOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
          description: 固定班单节默认教练费；为空时不变更标准
        private_lesson_fee:
          anyOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
          description: 私教单节默认教练费；为空时不变更标准
        rate_effective_from:
          type:
            - string
            - "null"
          format: date
    CoachUpdate:
      allOf:
        - $ref: "#/components/schemas/CoachWrite"
        - type: object
          required:
            - version
          properties:
            version:
              type: integer
              minimum: 1
    Coach:
      allOf:
        - $ref: "#/components/schemas/IdentifierVersion"
        - $ref: "#/components/schemas/CoachWrite"
        - type: object
          required:
            - is_active
            - fixed_class_fee
            - private_lesson_fee
          properties:
            is_active:
              type: boolean
            fixed_class_fee:
              $ref: "#/components/schemas/Money"
            private_lesson_fee:
              $ref: "#/components/schemas/Money"
            fixed_class_fee_effective_from:
              type:
                - string
                - "null"
              format: date
            private_lesson_fee_effective_from:
              type:
                - string
                - "null"
              format: date
    CoachRate:
      allOf:
        - $ref: "#/components/schemas/IdentifierVersion"
        - type: object
          required:
            - business_type
            - amount
            - effective_from
          properties:
            business_type:
              type: string
              enum:
                - fixed_class
                - private_lesson
            amount:
              $ref: "#/components/schemas/Money"
            effective_from:
              type: string
              format: date
            effective_to:
              type:
                - string
                - "null"
              format: date
    VenueSettingsWrite:
      type: object
      required:
        - name
        - timezone
        - weekday_open_time
        - weekday_close_time
        - weekend_open_time
        - weekend_close_time
        - version
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        timezone:
          type: string
          examples:
            - Asia/Shanghai
        weekday_open_time:
          type: string
          format: time
        weekday_close_time:
          type: string
          format: time
        weekend_open_time:
          type: string
          format: time
        weekend_close_time:
          type: string
          format: time
        version:
          type: integer
          minimum: 1
    VenueSettings:
      allOf:
        - $ref: "#/components/schemas/IdentifierVersion"
        - $ref: "#/components/schemas/VenueSettingsWrite"
    StudentWrite:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        gender:
          type:
            - string
            - "null"
        birth_date:
          type:
            - string
            - "null"
          format: date
        phone:
          type:
            - string
            - "null"
        level_or_tags:
          type: array
          items:
            type: string
        guardian_ids:
          type: array
          items:
            type: string
            format: uuid
        notes:
          type:
            - string
            - "null"
    Student:
      allOf:
        - $ref: "#/components/schemas/IdentifierVersion"
        - $ref: "#/components/schemas/StudentWrite"
        - type: object
          required:
            - status
          properties:
            status:
              type: string
              enum:
                - active
                - inactive
    StudentPage:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        next_cursor:
          type:
            - string
            - "null"
...